pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "lizard-fpv"
version = "0.1.0"
description = "LIZARD – Fraud Pattern Visualizer (timeline, map, link graph) with AI-assisted schema mapping"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "LIZARD Team", email = "mehdi.boustala@amadeus.com" }
]

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "aiosqlite>=0.20.0",
  "alembic>=1.13.2",
  "python-dotenv>=1.0.1",
  "pandas>=2.2.2",
  "numpy>=2.1.0",
  "structlog>=24.1.0",
  "prometheus-client>=0.20.0",
  "typer>=0.12.5",
  "pyyaml>=6.0.2",
  "python-json-logger>=2.0.7",
  "orjson>=3.10.7",
  "fastapi-cli>=0.0.5",
  "greenlet>=3.0.3",
  "pyarrow>=17.0.0",
  "python-multipart>=0.0.9",
  "scikit-learn>=1.5.0",
  "networkx>=3.2",
  "httpx>=0.27.2",
]

[project.optional-dependencies]
ai = [
  "openai>=1.40.0"
]
test = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.2",
  "anyio>=4.4.0",
  "mypy>=1.11.2",
  "types-requests>=2.32.0.20240914",
  "pytest-asyncio>=0.24.0",
]
dev = [
  "ruff>=0.6.3",
  "black>=24.8.0",
  "pre-commit>=3.8.0"
]
cloud = [
  "azure-identity>=1.17.0",
  "azure-storage-blob>=12.22.0",
  "databricks-sdk>=0.30.0",
]

# REMOVED broken script entry point
# [project.scripts]
# lizard-csv = "connectors.csv.loader:app"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
src = ["app", "domain", "connectors", "mapping", "analytics", "ui", "scripts", "cloud"]
select = ["E","F","I","UP","B","N","S","C4","PL","RUF"]
ignore = ["E203","E266","E501"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = false

[tool.hatch.build.targets.wheel]
packages = ["app","domain","connectors","mapping","analytics","ui","scripts","cloud"]